Adicionando Migrations

+++ SafeZoneModel/LeituraModel.cs
@@ -4,9 +4,9 @@
         public class LeituraModel
         {
             public int Id { get; set; }                      
-            public DateTime DataHora { get; set; }           
-            public double Temperatura { get; set; }          
-            public double Umidade { get; set; }                      
+            public DateTime DataHora { get; set; }
+            public decimal Temperatura { get; set; }
+            public decimal Umidade { get; set; }
         }
     }
